refactor: 선택된 값에 따라 보여주도록 render변경
