+++ src/event/triggerClickEvent.js
@@ -1,6 +1,6 @@
 import { removeTodoItem } from '../todoList/removeTodoItem.js';
-import { setViewTodo } from '../todoList/setViewTodo.js';
 import { toggleTodoItem } from '../todoList/toggleTodoItem.js';
+import { render } from '../render/render.js';
 
 export const triggerClickEvent = ({ target }) => {
   if (target.className === 'toggle') {
@@ -10,6 +10,6 @@ export const triggerClickEvent = ({ target }) => {
     removeTodoItem(target.closest('li'));
   }
   if (target.tagName === 'A') {
-    setViewTodo(target);
+    render(target);
   }
 };
