+++ src/todoList/addTodoItem.js
@@ -1,4 +1,4 @@
-import { render } from './render.js';
+import { render } from '../render/render.js';
 
 export const addTodoItem = (text) => {
   const todos = JSON.parse(localStorage.getItem('todos'));
